+++ model/strawberry/output/observation/statistics/component/__init__.py
@@ -1,0 +1,14 @@
+"""Wikibase Statistics Observation Components"""
+
+from model.strawberry.output.observation.statistics.component.edit_stats import (
+    WikibaseStatisticsEditsObservationStrawberryModel,
+)
+from model.strawberry.output.observation.statistics.component.files_stats import (
+    WikibaseStatisticsFilesObservationStrawberryModel,
+)
+from model.strawberry.output.observation.statistics.component.pages_stats import (
+    WikibaseStatisticsPagesObservationStrawberryModel,
+)
+from model.strawberry.output.observation.statistics.component.users_stats import (
+    WikibaseStatisticsUsersObservationStrawberryModel,
+)
