+++ model/strawberry/output/observation/statistics/wikibase_statistics_aggregate.py
@@ -34,6 +34,14 @@ class WikibaseStatisticsAggregateStrawberryModel:  # pylint: disable=too-many-in
             return None
         return self.content_page_word_count_total / self.content_pages
 
+    @strawberry.field(description="Average Edits per Page")
+    def edits_per_page_avg(self) -> Optional[float]:
+        """Average Edits per Page"""
+
+        if self.total_pages == 0:
+            return None
+        return self.total_edits / self.total_pages
+
     def __init__(  # pylint: disable=too-many-arguments
         self,
         total_pages: int,
